+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_session_receiver_async.py
@@ -90,7 +90,7 @@ class ServiceBusSessionReceiver(ServiceBusReceiver, SessionReceiverMixin):
         self._session = ServiceBusSession(self._session_id, self, self._config.encoding)
 
     @classmethod
-    def from_connection_string(
+    def _from_connection_string(
         cls,
         conn_str: str,
         **kwargs: Any
@@ -146,7 +146,7 @@ class ServiceBusSessionReceiver(ServiceBusReceiver, SessionReceiverMixin):
                 :caption: Create a new instance of the ServiceBusReceiver from connection string.
 
         """
-        return super(ServiceBusSessionReceiver, cls).from_connection_string(conn_str, **kwargs)  # type: ignore
+        return super(ServiceBusSessionReceiver, cls)._from_connection_string(conn_str, **kwargs)  # type: ignore
 
     @property
     def session(self):