+++ sdk/servicebus/azure-servicebus/tests/async_tests/test_queues_async.py
@@ -415,8 +415,8 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                     print_message(_logger, message)
                     assert message.dead_letter_reason == 'Testing reason'
                     assert message.dead_letter_error_description == 'Testing description'
-                    assert message.properties[b'DeadLetterReason'] == b'Testing reason'
-                    assert message.properties[b'DeadLetterErrorDescription'] == b'Testing description'
+                    assert message.application_properties[b'DeadLetterReason'] == b'Testing reason'
+                    assert message.application_properties[b'DeadLetterErrorDescription'] == b'Testing description'
                     await message.complete()
             assert count == 10
 
@@ -532,8 +532,8 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                     count += 1
                     assert message.dead_letter_reason == 'Testing reason'
                     assert message.dead_letter_error_description == 'Testing description'
-                    assert message.properties[b'DeadLetterReason'] == b'Testing reason'
-                    assert message.properties[b'DeadLetterErrorDescription'] == b'Testing description'
+                    assert message.application_properties[b'DeadLetterReason'] == b'Testing reason'
+                    assert message.application_properties[b'DeadLetterErrorDescription'] == b'Testing description'
                 assert count == 10
 
     @pytest.mark.liveTest
@@ -574,8 +574,8 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                     print_message(_logger, message)
                     assert message.dead_letter_reason == 'Testing reason'
                     assert message.dead_letter_error_description == 'Testing description'
-                    assert message.properties[b'DeadLetterReason'] == b'Testing reason'
-                    assert message.properties[b'DeadLetterErrorDescription'] == b'Testing description'
+                    assert message.application_properties[b'DeadLetterReason'] == b'Testing reason'
+                    assert message.application_properties[b'DeadLetterErrorDescription'] == b'Testing description'
                     await message.complete()
                     count += 1
             assert count == 10
@@ -1249,7 +1249,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                 for i in range(20):
                     yield Message(
                         body="Message no. {}".format(i),
-                        label='1st'
+                        subject='1st'
                     )
 
             sender = sb_client.get_queue_sender(servicebus_queue.name)
@@ -1275,12 +1275,12 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                     receive_counter += 1
                     for message in messages:
                         print_message(_logger, message)
-                        if message.label == '1st':
+                        if message.subject == '1st':
                             message_1st_received_cnt += 1
                             await message.complete()
-                            message.label = '2nd'
+                            message.subject = '2nd'
                             await sender.send_messages(message)  # resending received message
-                        elif message.label == '2nd':
+                        elif message.subject == '2nd':
                             message_2nd_received_cnt += 1
                             await message.complete()