+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -15,6 +15,14 @@
 * Updated uAMQP dependency to 1.2.11.
   - Fixed bug where amqp message `footer` and `delivery_annotation` were not encoded into the outgoing payload.
 
+**Breaking Changes**
+
+* `amqp_message` has been renamed to `amqp_annotated_message` for cross-sdk consistency.
+* All `name` parameters in `ServiceBusAdministrationClient` are now precisely specified ala `queue_name` or `rule_name`
+* `via_partition_key` is no longer exposed, this is pending a full implementation of transactions as it has no external use.  If needed, the underlying value can still be accessed in `Message.amqp_annotated_message.annotations`.
+* `Message.properties` has been renamed to `Message.application_properties` for consistency with service verbiage.
+* Sub-client (`ServiceBusSender` and `ServiceBusReceiver`) `from_connection_string` initializers have been made internal until needed.  Clients should be initialized from root `ServiceBusClient`.
+
 ## 7.0.0b7 (2020-10-05)
 
 **Breaking Changes**